+++ source/lambda/executor/main.py
@@ -11,14 +11,14 @@ import asyncio
 import math
 
 import boto3
-from langchain.callbacks.base import BaseCallbackHandler
-from langchain.globals import set_verbose
-# from langchain.llms import OpenAI
-from langchain.output_parsers import PydanticOutputParser
-from langchain.prompts import ChatPromptTemplate
-from langchain.pydantic_v1 import BaseModel, Field, validator
-from langchain.retrievers import ContextualCompressionRetriever
-from langchain.retrievers.document_compressors import CohereRerank
+# from langchain.callbacks.base import BaseCallbackHandler
+# from langchain.globals import set_verbose
+# # from langchain.llms import OpenAI
+# from langchain.output_parsers import PydanticOutputParser
+# from langchain.prompts import ChatPromptTemplate
+# from langchain.pydantic_v1 import BaseModel, Field, validator
+# from langchain.retrievers import ContextualCompressionRetriever
+# from langchain.retrievers.document_compressors import CohereRerank
 
 # from llm_utils import CustomLLM
 from langchain.retrievers.merger_retriever import MergerRetriever
@@ -29,48 +29,56 @@ from langchain.schema.runnable import (
     RunnableParallel,
     RunnablePassthrough,
 )
-from langchain.schema.messages import (
-    HumanMessage,AIMessage,SystemMessage
-)
+# from langchain.schema.messages import (
+#     HumanMessage,AIMessage,SystemMessage
+# )
 # from langchain.memory import ConversationSummaryMemory, ChatMessageHistory
 # from langchain.utilities import GoogleSearchAPIWrapper
-from dateutil import parser
+# from dateutil import parser
 
-from utils.reranker import BGEReranker, MergeReranker
+# from utils.reranker import BGEReranker, MergeReranker
 from utils.retriever import (
     QueryDocumentRetriever,
     QueryQuestionRetriever,
     index_results_format,
 )
-from langchain.retrievers.multi_query import MultiQueryRetriever
-from utils.logger_utils import logger,opensearch_logger,boto3_logger
+# from langchain.retrievers.multi_query import MultiQueryRetriever
+from utils.logger_utils import logger
 
 
 from utils.aos_utils import LLMBotOpenSearchClient
-from utils.constant import IntentType, Type
-from utils.ddb_utils import DynamoDBChatMessageHistory,filter_chat_history_by_time
-from utils.intent_utils import auto_intention_recoginition_chain
+from utils.constant import Type
+from utils.ddb_utils import DynamoDBChatMessageHistory
+# from utils.intent_utils import auto_intention_recoginition_chain
 # from langchain_utils import create_identity_lambda
 
 # from llm_utils import generate as llm_generate
-from utils.llm_utils import get_llm_chain,get_llm_model
-from utils.llmbot_utils import (
-    # QueryType,
-    combine_recalls,
-    # concat_recall_knowledge,
-    # process_input_messages,
-)
-from utils.time_utils import timeit
-from utils.preprocess_utils import run_preprocess
+# from utils.llm_utils import get_llm_chain,get_llm_model
+# from utils.llmbot_utils import (
+#     # QueryType,
+#     combine_recalls,
+#     # concat_recall_knowledge,
+#     # process_input_messages,
+# )
+# from utils.time_utils import timeit
+# from utils.preprocess_utils import run_preprocess
 from utils.response_utils import process_response
-from utils.sm_utils import SagemakerEndpointVectorOrCross
-from utils.constant import Type,IntentType
-from utils.intent_utils import auto_intention_recoginition_chain
-from utils.langchain_utils import add_key_to_debug,chain_logger
-from utils.query_process_utils import get_query_process_chain
+# from utils.sm_utils import SagemakerEndpointVectorOrCross
+from utils.constant import Type
+# from utils.intent_utils import auto_intention_recoginition_chain
+# from utils.langchain_utils import add_key_to_debug,chain_logger
+# from utils.query_process_utils import get_query_process_chain
 import utils.parse_config as parse_config
 from utils.serialization_utils import JSONEncoder
-from utils.constant import MKT_CONVERSATION_SUMMARY_TYPE
+from utils.executor_entries import (
+    market_chain_entry,
+    market_conversation_summary_entry,
+    main_chain_entry,
+    main_qd_retriever_entry,
+    main_qq_retriever_entry,
+    get_retriever_response
+)
+# from utils.constant import MKT_CONVERSATION_SUMMARY_TYPE
 
 region = os.environ["AWS_REGION"]
 embedding_endpoint = os.environ.get("embedding_endpoint", "")
@@ -85,8 +93,8 @@ aos_ug_index = os.environ.get("aos_ug_index", "")
 llm_endpoint = os.environ.get("llm_endpoint", "")
 chat_session_table = os.environ.get("chat_session_table", "")
 websocket_url = os.environ.get("websocket_url", "")
-sm_client = boto3.client("sagemaker-runtime")
-aos_client = LLMBotOpenSearchClient(aos_endpoint)
+# sm_client = boto3.client("sagemaker-runtime")
+# aos_client = LLMBotOpenSearchClient(aos_endpoint)
 ws_client = None
 
 # get aos_index_dict
@@ -123,461 +131,461 @@ def handle_error(func):
     return wrapper
 
 
-def get_faq_answer(source, index_name):
-    opensearch_query_response = aos_client.search(
-        index_name=index_name,
-        query_type="basic",
-        query_term=source,
-        field="metadata.source",
-    )
-    for r in opensearch_query_response["hits"]["hits"]:
-        if r["_source"]["metadata"]["field"] == "answer":
-            return r["_source"]["content"]
-    return ""
+# def get_faq_answer(source, index_name):
+#     opensearch_query_response = aos_client.search(
+#         index_name=index_name,
+#         query_type="basic",
+#         query_term=source,
+#         field="metadata.source",
+#     )
+#     for r in opensearch_query_response["hits"]["hits"]:
+#         if r["_source"]["metadata"]["field"] == "answer":
+#             return r["_source"]["content"]
+#     return ""
 
 
-def get_faq_content(source, index_name):
-    opensearch_query_response = aos_client.search(
-        index_name=index_name,
-        query_type="basic",
-        query_term=source,
-        field="metadata.source",
-    )
-    for r in opensearch_query_response["hits"]["hits"]:
-        if r["_source"]["metadata"]["field"] == "all_text":
-            return r["_source"]["content"]
-    return ""
+# def get_faq_content(source, index_name):
+#     opensearch_query_response = aos_client.search(
+#         index_name=index_name,
+#         query_type="basic",
+#         query_term=source,
+#         field="metadata.source",
+#     )
+#     for r in opensearch_query_response["hits"]["hits"]:
+#         if r["_source"]["metadata"]["field"] == "all_text":
+#             return r["_source"]["content"]
+#     return ""
 
 
-def organize_faq_results(response, index_name):
-    """
-    Organize results from aos response
+# def organize_faq_results(response, index_name):
+#     """
+#     Organize results from aos response
 
-    :param query_type: query type
-    :param response: aos response json
-    """
-    results = []
-    if not response:
-        return results
-    aos_hits = response["hits"]["hits"]
-    for aos_hit in aos_hits:
-        result = {}
-        try:
-            result["source"] = aos_hit["_source"]["metadata"]["source"]
-            result["score"] = aos_hit["_score"]
-            result["detail"] = aos_hit["_source"]
-            result["content"] = aos_hit["_source"]["content"]
-            result["answer"] = get_faq_answer(result["source"], index_name)
-            result["doc"] = get_faq_content(result["source"], index_name)
-        except:
-            logger.info("index_error")
-            logger.info(aos_hit["_source"])
-            continue
-        # result.update(aos_hit["_source"])
-        results.append(result)
-    return results
-
-
-def get_ug_content(source, index_name):
-    opensearch_query_response = aos_client.search(
-        index_name=index_name,
-        query_type="basic",
-        query_term=source,
-        field="metadata.source",
-        size=100,
-    )
-    for r in opensearch_query_response["hits"]["hits"]:
-        if r["_source"]["metadata"]["field"] == "all_text":
-            return r["_source"]["content"]
-    return ""
+#     :param query_type: query type
+#     :param response: aos response json
+#     """
+#     results = []
+#     if not response:
+#         return results
+#     aos_hits = response["hits"]["hits"]
+#     for aos_hit in aos_hits:
+#         result = {}
+#         try:
+#             result["source"] = aos_hit["_source"]["metadata"]["source"]
+#             result["score"] = aos_hit["_score"]
+#             result["detail"] = aos_hit["_source"]
+#             result["content"] = aos_hit["_source"]["content"]
+#             result["answer"] = get_faq_answer(result["source"], index_name)
+#             result["doc"] = get_faq_content(result["source"], index_name)
+#         except:
+#             logger.info("index_error")
+#             logger.info(aos_hit["_source"])
+#             continue
+#         # result.update(aos_hit["_source"])
+#         results.append(result)
+#     return results
 
 
-def organize_ug_results(response, index_name):
-    """
-    Organize results from aos response
-
-    :param query_type: query type
-    :param response: aos response json
-    """
-    results = []
-    aos_hits = response["hits"]["hits"]
-    for aos_hit in aos_hits:
-        result = {}
-        result["source"] = aos_hit["_source"]["metadata"]["source"]
-        result["score"] = aos_hit["_score"]
-        result["detail"] = aos_hit["_source"]
-        result["content"] = aos_hit["_source"]["content"]
-        result["doc"] = get_ug_content(result["source"], index_name)
-        # result.update(aos_hit["_source"])
-        results.append(result)
-    return results
-
-def remove_redundancy_debug_info(results):
-    filtered_results = copy.deepcopy(results)
-    for result in filtered_results:
-        for field in list(result["detail"].keys()):
-            if field.endswith("embedding") or field.startswith("vector"):
-                del result["detail"][field]
-    return filtered_results
-
-def parse_query(
-    query_input: str,
-    history: list,
-    zh_embedding_model_endpoint: str,
-    en_embedding_model_endpoint: str,
-    debug_info: dict,
-):
-    start = time.time()
-    # concatenate query_input and history to unified prompt
-    query_knowledge = "".join([query_input] + [row[0] for row in history][::-1])
-
-    # get query embedding
-    parsed_query = run_preprocess(query_knowledge)
-    debug_info["query_parser_info"] = parsed_query
-    if parsed_query["query_lang"] == "zh":
-        parsed_query["zh_query"] = query_knowledge
-        parsed_query["en_query"] = parsed_query["translated_text"]
-    elif parsed_query["query_lang"] == "en":
-        parsed_query["zh_query"] = parsed_query["translated_text"]
-        parsed_query["en_query"] = query_knowledge
-    zh_query_similarity_embedding_prompt = parsed_query["zh_query"]
-    en_query_similarity_embedding_prompt = parsed_query["en_query"]
-    zh_query_relevance_embedding_prompt = (
-        "为这个句子生成表示以用于检索相关文章：" + parsed_query["zh_query"]
-    )
-    en_query_relevance_embedding_prompt = (
-        "Represent this sentence for searching relevant passages: "
-        + parsed_query["en_query"]
-    )
-    parsed_query["zh_query_similarity_embedding"] = SagemakerEndpointVectorOrCross(
-        prompt=zh_query_similarity_embedding_prompt,
-        endpoint_name=zh_embedding_model_endpoint,
-        region_name=region,
-        model_type="vector",
-        stop=None,
-    )
-    parsed_query["zh_query_relevance_embedding"] = SagemakerEndpointVectorOrCross(
-        prompt=zh_query_relevance_embedding_prompt,
-        endpoint_name=zh_embedding_model_endpoint,
-        region_name=region,
-        model_type="vector",
-        stop=None,
-    )
-    parsed_query["en_query_similarity_embedding"] = SagemakerEndpointVectorOrCross(
-        prompt=en_query_similarity_embedding_prompt,
-        endpoint_name=en_embedding_model_endpoint,
-        region_name=region,
-        model_type="vector",
-        stop=None,
-    )
-    parsed_query["en_query_relevance_embedding"] = SagemakerEndpointVectorOrCross(
-        prompt=en_query_relevance_embedding_prompt,
-        endpoint_name=en_embedding_model_endpoint,
-        region_name=region,
-        model_type="vector",
-        stop=None,
-    )
-    elpase_time = time.time() - start
-    logger.info(f"runing time of parse query: {elpase_time}s seconds")
-    return parsed_query
+# def get_ug_content(source, index_name):
+#     opensearch_query_response = aos_client.search(
+#         index_name=index_name,
+#         query_type="basic",
+#         query_term=source,
+#         field="metadata.source",
+#         size=100,
+#     )
+#     for r in opensearch_query_response["hits"]["hits"]:
+#         if r["_source"]["metadata"]["field"] == "all_text":
+#             return r["_source"]["content"]
+#     return ""
 
 
-def q_q_match(parsed_query, debug_info):
-    start = time.time()
-    opensearch_knn_results = []
-    opensearch_knn_response = aos_client.search(
-        index_name=aos_faq_index,
-        query_type="knn",
-        query_term=parsed_query["zh_query_similarity_embedding"],
-        field="embedding",
-        size=2,
-    )
-    opensearch_knn_results.extend(
-        organize_faq_results(opensearch_knn_response, aos_faq_index)
-    )
-    opensearch_knn_response = aos_client.search(
-        index_name=aos_faq_index,
-        query_type="knn",
-        query_term=parsed_query["en_query_similarity_embedding"],
-        field="embedding",
-        size=2,
-    )
-    opensearch_knn_results.extend(
-        organize_faq_results(opensearch_knn_response, aos_faq_index)
-    )
-    # logger.info(json.dumps(opensearch_knn_response, ensure_ascii=False))
-    elpase_time = time.time() - start
-    logger.info(f"runing time of opensearch_knn : {elpase_time}s seconds")
-    answer = None
-    sources = None
-    if len(opensearch_knn_results) > 0:
-        debug_info["q_q_match_info"] = remove_redundancy_debug_info(
-            opensearch_knn_results[:3]
-        )
-        if opensearch_knn_results[0]["score"] >= 0.9:
-            source = opensearch_knn_results[0]["source"]
-            answer = opensearch_knn_results[0]["answer"]
-            sources = [source]
-            return answer, sources
-    return answer, sources
+# def organize_ug_results(response, index_name):
+#     """
+#     Organize results from aos response
+
+#     :param query_type: query type
+#     :param response: aos response json
+#     """
+#     results = []
+#     aos_hits = response["hits"]["hits"]
+#     for aos_hit in aos_hits:
+#         result = {}
+#         result["source"] = aos_hit["_source"]["metadata"]["source"]
+#         result["score"] = aos_hit["_score"]
+#         result["detail"] = aos_hit["_source"]
+#         result["content"] = aos_hit["_source"]["content"]
+#         result["doc"] = get_ug_content(result["source"], index_name)
+#         # result.update(aos_hit["_source"])
+#         results.append(result)
+#     return results
+
+# def remove_redundancy_debug_info(results):
+#     filtered_results = copy.deepcopy(results)
+#     for result in filtered_results:
+#         for field in list(result["detail"].keys()):
+#             if field.endswith("embedding") or field.startswith("vector"):
+#                 del result["detail"][field]
+#     return filtered_results
+
+# def parse_query(
+#     query_input: str,
+#     history: list,
+#     zh_embedding_model_endpoint: str,
+#     en_embedding_model_endpoint: str,
+#     debug_info: dict,
+# ):
+#     start = time.time()
+#     # concatenate query_input and history to unified prompt
+#     query_knowledge = "".join([query_input] + [row[0] for row in history][::-1])
+
+#     # get query embedding
+#     parsed_query = run_preprocess(query_knowledge)
+#     debug_info["query_parser_info"] = parsed_query
+#     if parsed_query["query_lang"] == "zh":
+#         parsed_query["zh_query"] = query_knowledge
+#         parsed_query["en_query"] = parsed_query["translated_text"]
+#     elif parsed_query["query_lang"] == "en":
+#         parsed_query["zh_query"] = parsed_query["translated_text"]
+#         parsed_query["en_query"] = query_knowledge
+#     zh_query_similarity_embedding_prompt = parsed_query["zh_query"]
+#     en_query_similarity_embedding_prompt = parsed_query["en_query"]
+#     zh_query_relevance_embedding_prompt = (
+#         "为这个句子生成表示以用于检索相关文章：" + parsed_query["zh_query"]
+#     )
+#     en_query_relevance_embedding_prompt = (
+#         "Represent this sentence for searching relevant passages: "
+#         + parsed_query["en_query"]
+#     )
+#     parsed_query["zh_query_similarity_embedding"] = SagemakerEndpointVectorOrCross(
+#         prompt=zh_query_similarity_embedding_prompt,
+#         endpoint_name=zh_embedding_model_endpoint,
+#         region_name=region,
+#         model_type="vector",
+#         stop=None,
+#     )
+#     parsed_query["zh_query_relevance_embedding"] = SagemakerEndpointVectorOrCross(
+#         prompt=zh_query_relevance_embedding_prompt,
+#         endpoint_name=zh_embedding_model_endpoint,
+#         region_name=region,
+#         model_type="vector",
+#         stop=None,
+#     )
+#     parsed_query["en_query_similarity_embedding"] = SagemakerEndpointVectorOrCross(
+#         prompt=en_query_similarity_embedding_prompt,
+#         endpoint_name=en_embedding_model_endpoint,
+#         region_name=region,
+#         model_type="vector",
+#         stop=None,
+#     )
+#     parsed_query["en_query_relevance_embedding"] = SagemakerEndpointVectorOrCross(
+#         prompt=en_query_relevance_embedding_prompt,
+#         endpoint_name=en_embedding_model_endpoint,
+#         region_name=region,
+#         model_type="vector",
+#         stop=None,
+#     )
+#     elpase_time = time.time() - start
+#     logger.info(f"runing time of parse query: {elpase_time}s seconds")
+#     return parsed_query
 
 
-def get_relevant_documents_dgr(
-    parsed_query,
-    rerank_model_endpoint: str,
-    aos_faq_index: str,
-    aos_ug_index: str,
-    debug_info,
-):
-    # 1. get AOS knn recall
-    faq_result_num = 2
-    ug_result_num = 20
-    start = time.time()
-    opensearch_knn_results = []
-    opensearch_knn_response = aos_client.search(
-        index_name=aos_faq_index,
-        query_type="knn",
-        query_term=parsed_query["zh_query_relevance_embedding"],
-        field="embedding",
-        size=faq_result_num,
-    )
-    opensearch_knn_results.extend(
-        organize_faq_results(opensearch_knn_response, aos_faq_index)[:faq_result_num]
-    )
-    opensearch_knn_response = aos_client.search(
-        index_name=aos_faq_index,
-        query_type="knn",
-        query_term=parsed_query["en_query_relevance_embedding"],
-        field="embedding",
-        size=faq_result_num,
-    )
-    opensearch_knn_results.extend(
-        organize_faq_results(opensearch_knn_response, aos_faq_index)[:faq_result_num]
-    )
-    # logger.info(json.dumps(opensearch_knn_response, ensure_ascii=False))
-    faq_recall_end_time = time.time()
-    elpase_time = faq_recall_end_time - start
-    logger.info(f"runing time of faq recall : {elpase_time}s seconds")
-    filter = None
-    if parsed_query["is_api_query"]:
-        filter = [{"term": {"metadata.is_api": True}}]
-
-    opensearch_knn_response = aos_client.search(
-        index_name=aos_ug_index,
-        query_type="knn",
-        query_term=parsed_query["zh_query_relevance_embedding"],
-        field="embedding",
-        filter=filter,
-        size=ug_result_num,
-    )
-    opensearch_knn_results.extend(
-        organize_ug_results(opensearch_knn_response, aos_ug_index)[:ug_result_num]
-    )
-    opensearch_knn_response = aos_client.search(
-        index_name=aos_ug_index,
-        query_type="knn",
-        query_term=parsed_query["en_query_relevance_embedding"],
-        field="embedding",
-        filter=filter,
-        size=ug_result_num,
-    )
-    opensearch_knn_results.extend(
-        organize_ug_results(opensearch_knn_response, aos_ug_index)[:ug_result_num]
-    )
-
-    debug_info["knowledge_qa_knn_recall"] = remove_redundancy_debug_info(
-        opensearch_knn_results
-    )
-    ug_recall_end_time = time.time()
-    elpase_time = ug_recall_end_time - faq_recall_end_time
-    logger.info(f"runing time of ug recall: {elpase_time}s seconds")
-
-    # 2. get AOS invertedIndex recall
-    opensearch_query_results = []
-
-    # 3. combine these two opensearch_knn_response and opensearch_query_response
-    recall_knowledge = combine_recalls(opensearch_knn_results, opensearch_query_results)
-
-    rerank_pair = []
-    for knowledge in recall_knowledge:
-        # rerank_pair.append([parsed_query["query"], knowledge["content"]][:1024])
-        rerank_pair.append(
-            [parsed_query["en_query"], knowledge["content"]][: 1024 * 10]
-        )
-    en_score_list = json.loads(
-        SagemakerEndpointVectorOrCross(
-            prompt=json.dumps(rerank_pair),
-            endpoint_name=rerank_model_endpoint,
-            region_name=region,
-            model_type="rerank",
-            stop=None,
-        )
-    )
-    rerank_pair = []
-    for knowledge in recall_knowledge:
-        # rerank_pair.append([parsed_query["query"], knowledge["content"]][:1024])
-        rerank_pair.append(
-            [parsed_query["zh_query"], knowledge["content"]][: 1024 * 10]
-        )
-    zh_score_list = json.loads(
-        SagemakerEndpointVectorOrCross(
-            prompt=json.dumps(rerank_pair),
-            endpoint_name=rerank_model_endpoint,
-            region_name=region,
-            model_type="rerank",
-            stop=None,
-        )
-    )
-    rerank_knowledge = []
-    for knowledge, score in zip(recall_knowledge, zh_score_list):
-        # if score > 0:
-        new_knowledge = knowledge.copy()
-        new_knowledge["rerank_score"] = score
-        rerank_knowledge.append(new_knowledge)
-    for knowledge, score in zip(recall_knowledge, en_score_list):
-        # if score > 0:
-        new_knowledge = knowledge.copy()
-        new_knowledge["rerank_score"] = score
-        rerank_knowledge.append(new_knowledge)
-    rerank_knowledge.sort(key=lambda x: x["rerank_score"], reverse=True)
-    debug_info["knowledge_qa_rerank"] = rerank_knowledge
-
-    rerank_end_time = time.time()
-    elpase_time = rerank_end_time - ug_recall_end_time
-    logger.info(f"runing time of rerank: {elpase_time}s seconds")
-
-    return rerank_knowledge
+# def q_q_match(parsed_query, debug_info):
+#     start = time.time()
+#     opensearch_knn_results = []
+#     opensearch_knn_response = aos_client.search(
+#         index_name=aos_faq_index,
+#         query_type="knn",
+#         query_term=parsed_query["zh_query_similarity_embedding"],
+#         field="embedding",
+#         size=2,
+#     )
+#     opensearch_knn_results.extend(
+#         organize_faq_results(opensearch_knn_response, aos_faq_index)
+#     )
+#     opensearch_knn_response = aos_client.search(
+#         index_name=aos_faq_index,
+#         query_type="knn",
+#         query_term=parsed_query["en_query_similarity_embedding"],
+#         field="embedding",
+#         size=2,
+#     )
+#     opensearch_knn_results.extend(
+#         organize_faq_results(opensearch_knn_response, aos_faq_index)
+#     )
+#     # logger.info(json.dumps(opensearch_knn_response, ensure_ascii=False))
+#     elpase_time = time.time() - start
+#     logger.info(f"runing time of opensearch_knn : {elpase_time}s seconds")
+#     answer = None
+#     sources = None
+#     if len(opensearch_knn_results) > 0:
+#         debug_info["q_q_match_info"] = remove_redundancy_debug_info(
+#             opensearch_knn_results[:3]
+#         )
+#         if opensearch_knn_results[0]["score"] >= 0.9:
+#             source = opensearch_knn_results[0]["source"]
+#             answer = opensearch_knn_results[0]["answer"]
+#             sources = [source]
+#             return answer, sources
+#     return answer, sources
 
 
-def dgr_entry(
-    session_id: str,
-    query_input: str,
-    history: list,
-    zh_embedding_model_endpoint: str,
-    en_embedding_model_endpoint: str,
-    cross_model_endpoint: str,
-    rerank_model_endpoint: str,
-    llm_model_endpoint: str,
-    aos_faq_index: str,
-    aos_ug_index: str,
-    enable_knowledge_qa: bool,
-    temperature: float,
-    enable_q_q_match: bool,
-    llm_model_id=None,
-    stream=False,
-):
-    """
-    Entry point for the Lambda function.
+# def get_relevant_documents_dgr(
+#     parsed_query,
+#     rerank_model_endpoint: str,
+#     aos_faq_index: str,
+#     aos_ug_index: str,
+#     debug_info,
+# ):
+#     # 1. get AOS knn recall
+#     faq_result_num = 2
+#     ug_result_num = 20
+#     start = time.time()
+#     opensearch_knn_results = []
+#     opensearch_knn_response = aos_client.search(
+#         index_name=aos_faq_index,
+#         query_type="knn",
+#         query_term=parsed_query["zh_query_relevance_embedding"],
+#         field="embedding",
+#         size=faq_result_num,
+#     )
+#     opensearch_knn_results.extend(
+#         organize_faq_results(opensearch_knn_response, aos_faq_index)[:faq_result_num]
+#     )
+#     opensearch_knn_response = aos_client.search(
+#         index_name=aos_faq_index,
+#         query_type="knn",
+#         query_term=parsed_query["en_query_relevance_embedding"],
+#         field="embedding",
+#         size=faq_result_num,
+#     )
+#     opensearch_knn_results.extend(
+#         organize_faq_results(opensearch_knn_response, aos_faq_index)[:faq_result_num]
+#     )
+#     # logger.info(json.dumps(opensearch_knn_response, ensure_ascii=False))
+#     faq_recall_end_time = time.time()
+#     elpase_time = faq_recall_end_time - start
+#     logger.info(f"runing time of faq recall : {elpase_time}s seconds")
+#     filter = None
+#     if parsed_query["is_api_query"]:
+#         filter = [{"term": {"metadata.is_api": True}}]
 
-    :param session_id: The ID of the session.
-    :param query_input: The query input.
-    :param history: The history of the conversation.
-    :param embedding_model_endpoint: The endpoint of the embedding model.
-    :param cross_model_endpoint: The endpoint of the cross model.
-    :param llm_model_endpoint: The endpoint of the language model.
-    :param llm_model_name: The name of the language model.
-    :param aos_faq_index: The faq index of the AOS engine.
-    :param aos_ug_index: The ug index of the AOS engine.
-    :param enable_knowledge_qa: Whether to enable knowledge QA.
-    :param temperature: The temperature of the language model.
-    :param stream(Bool): Whether to use llm stream decoding output.
+#     opensearch_knn_response = aos_client.search(
+#         index_name=aos_ug_index,
+#         query_type="knn",
+#         query_term=parsed_query["zh_query_relevance_embedding"],
+#         field="embedding",
+#         filter=filter,
+#         size=ug_result_num,
+#     )
+#     opensearch_knn_results.extend(
+#         organize_ug_results(opensearch_knn_response, aos_ug_index)[:ug_result_num]
+#     )
+#     opensearch_knn_response = aos_client.search(
+#         index_name=aos_ug_index,
+#         query_type="knn",
+#         query_term=parsed_query["en_query_relevance_embedding"],
+#         field="embedding",
+#         filter=filter,
+#         size=ug_result_num,
+#     )
+#     opensearch_knn_results.extend(
+#         organize_ug_results(opensearch_knn_response, aos_ug_index)[:ug_result_num]
+#     )
 
-    return: answer(str)
-    """
-    debug_info = {
-        "query": query_input,
-        "query_parser_info": {},
-        "q_q_match_info": {},
-        "knowledge_qa_knn_recall": {},
-        "knowledge_qa_boolean_recall": {},
-        "knowledge_qa_combined_recall": {},
-        "knowledge_qa_cross_model_sort": {},
-        "knowledge_qa_llm": {},
-        "knowledge_qa_rerank": {},
-    }
-    contexts = []
-    sources = []
-    answer = ""
-    try:
-        # 1. parse query
-        parsed_query = parse_query(
-            query_input,
-            history,
-            zh_embedding_model_endpoint,
-            en_embedding_model_endpoint,
-            debug_info,
-        )
-        # 2. query question match
-        if enable_q_q_match:
-            answer, sources = q_q_match(parsed_query, debug_info)
-            if answer and sources:
-                return answer, sources, contexts, debug_info
-        # 3. recall and rerank
-        knowledges = get_relevant_documents_dgr(
-            parsed_query,
-            rerank_model_endpoint,
-            aos_faq_index,
-            aos_ug_index,
-            debug_info,
-        )
-        context_num = 6
-        sources = list(set([item["source"] for item in knowledges[:context_num]]))
-        contexts = knowledges[:context_num]
-        # 4. generate answer using question and recall_knowledge
-        parameters = {"temperature": temperature}
-        generate_input = dict(
-            model_id=llm_model_id,
-            query=query_input,
-            contexts=knowledges[:context_num],
-            history=history,
-            region_name=region,
-            model_kwargs=parameters,
-            context_num=context_num,
-            model_type="answer",
-            llm_model_endpoint=llm_model_endpoint,
-            stream=stream,
-        )
+#     debug_info["knowledge_qa_knn_recall"] = remove_redundancy_debug_info(
+#         opensearch_knn_results
+#     )
+#     ug_recall_end_time = time.time()
+#     elpase_time = ug_recall_end_time - faq_recall_end_time
+#     logger.info(f"runing time of ug recall: {elpase_time}s seconds")
 
-        llm_start_time = time.time()
-        llm_chain = get_rag_llm_chain(**generate_input)
-        llm_chain.invoke()
+#     # 2. get AOS invertedIndex recall
+#     opensearch_query_results = []
 
-        answer = llm_generate(**generate_input)
-        llm_end_time = time.time()
-        elpase_time = llm_end_time - llm_start_time
-        logger.info(f"runing time of llm: {elpase_time}s seconds")
-        # answer = ret["answer"]
-        debug_info["knowledge_qa_llm"] = answer
-    except Exception as e:
-        logger.info(f"Exception Query: {query_input}")
-        logger.info(f"{traceback.format_exc()}")
-        answer = ""
+#     # 3. combine these two opensearch_knn_response and opensearch_query_response
+#     recall_knowledge = combine_recalls(opensearch_knn_results, opensearch_query_results)
 
-    # 5. update_session
-    # start = time.time()
-    # update_session(session_id=session_id, chat_session_table=chat_session_table,
-    #                question=query_input, answer=answer, knowledge_sources=sources)
-    # elpase_time = time.time() - start
-    # logger.info(f'runing time of update_session : {elpase_time}s seconds')
+#     rerank_pair = []
+#     for knowledge in recall_knowledge:
+#         # rerank_pair.append([parsed_query["query"], knowledge["content"]][:1024])
+#         rerank_pair.append(
+#             [parsed_query["en_query"], knowledge["content"]][: 1024 * 10]
+#         )
+#     en_score_list = json.loads(
+#         SagemakerEndpointVectorOrCross(
+#             prompt=json.dumps(rerank_pair),
+#             endpoint_name=rerank_model_endpoint,
+#             region_name=region,
+#             model_type="rerank",
+#             stop=None,
+#         )
+#     )
+#     rerank_pair = []
+#     for knowledge in recall_knowledge:
+#         # rerank_pair.append([parsed_query["query"], knowledge["content"]][:1024])
+#         rerank_pair.append(
+#             [parsed_query["zh_query"], knowledge["content"]][: 1024 * 10]
+#         )
+#     zh_score_list = json.loads(
+#         SagemakerEndpointVectorOrCross(
+#             prompt=json.dumps(rerank_pair),
+#             endpoint_name=rerank_model_endpoint,
+#             region_name=region,
+#             model_type="rerank",
+#             stop=None,
+#         )
+#     )
+#     rerank_knowledge = []
+#     for knowledge, score in zip(recall_knowledge, zh_score_list):
+#         # if score > 0:
+#         new_knowledge = knowledge.copy()
+#         new_knowledge["rerank_score"] = score
+#         rerank_knowledge.append(new_knowledge)
+#     for knowledge, score in zip(recall_knowledge, en_score_list):
+#         # if score > 0:
+#         new_knowledge = knowledge.copy()
+#         new_knowledge["rerank_score"] = score
+#         rerank_knowledge.append(new_knowledge)
+#     rerank_knowledge.sort(key=lambda x: x["rerank_score"], reverse=True)
+#     debug_info["knowledge_qa_rerank"] = rerank_knowledge
 
-    return answer, sources, contexts, debug_info
+#     rerank_end_time = time.time()
+#     elpase_time = rerank_end_time - ug_recall_end_time
+#     logger.info(f"runing time of rerank: {elpase_time}s seconds")
+
+#     return rerank_knowledge
 
 
-def get_strict_qq_chain(strict_q_q_index):
-    def get_strict_qq_result(docs, threshold=0.7):
-        results = []
-        for doc in docs:
-            if doc.metadata["score"] < threshold:
-                break 
-            results.append(
-                {
-                    "score": doc.metadata["score"],
-                    "source": doc.metadata["source"],
-                    "answer": doc.metadata["answer"],
-                    "question": doc.metadata["question"],
-                }
-            )
-        return results
+# def dgr_entry(
+#     session_id: str,
+#     query_input: str,
+#     history: list,
+#     zh_embedding_model_endpoint: str,
+#     en_embedding_model_endpoint: str,
+#     cross_model_endpoint: str,
+#     rerank_model_endpoint: str,
+#     llm_model_endpoint: str,
+#     aos_faq_index: str,
+#     aos_ug_index: str,
+#     enable_knowledge_qa: bool,
+#     temperature: float,
+#     enable_q_q_match: bool,
+#     llm_model_id=None,
+#     stream=False,
+# ):
+#     """
+#     Entry point for the Lambda function.
+
+#     :param session_id: The ID of the session.
+#     :param query_input: The query input.
+#     :param history: The history of the conversation.
+#     :param embedding_model_endpoint: The endpoint of the embedding model.
+#     :param cross_model_endpoint: The endpoint of the cross model.
+#     :param llm_model_endpoint: The endpoint of the language model.
+#     :param llm_model_name: The name of the language model.
+#     :param aos_faq_index: The faq index of the AOS engine.
+#     :param aos_ug_index: The ug index of the AOS engine.
+#     :param enable_knowledge_qa: Whether to enable knowledge QA.
+#     :param temperature: The temperature of the language model.
+#     :param stream(Bool): Whether to use llm stream decoding output.
+
+#     return: answer(str)
+#     """
+#     debug_info = {
+#         "query": query_input,
+#         "query_parser_info": {},
+#         "q_q_match_info": {},
+#         "knowledge_qa_knn_recall": {},
+#         "knowledge_qa_boolean_recall": {},
+#         "knowledge_qa_combined_recall": {},
+#         "knowledge_qa_cross_model_sort": {},
+#         "knowledge_qa_llm": {},
+#         "knowledge_qa_rerank": {},
+#     }
+#     contexts = []
+#     sources = []
+#     answer = ""
+#     try:
+#         # 1. parse query
+#         parsed_query = parse_query(
+#             query_input,
+#             history,
+#             zh_embedding_model_endpoint,
+#             en_embedding_model_endpoint,
+#             debug_info,
+#         )
+#         # 2. query question match
+#         if enable_q_q_match:
+#             answer, sources = q_q_match(parsed_query, debug_info)
+#             if answer and sources:
+#                 return answer, sources, contexts, debug_info
+#         # 3. recall and rerank
+#         knowledges = get_relevant_documents_dgr(
+#             parsed_query,
+#             rerank_model_endpoint,
+#             aos_faq_index,
+#             aos_ug_index,
+#             debug_info,
+#         )
+#         context_num = 6
+#         sources = list(set([item["source"] for item in knowledges[:context_num]]))
+#         contexts = knowledges[:context_num]
+#         # 4. generate answer using question and recall_knowledge
+#         parameters = {"temperature": temperature}
+#         generate_input = dict(
+#             model_id=llm_model_id,
+#             query=query_input,
+#             contexts=knowledges[:context_num],
+#             history=history,
+#             region_name=region,
+#             model_kwargs=parameters,
+#             context_num=context_num,
+#             model_type="answer",
+#             llm_model_endpoint=llm_model_endpoint,
+#             stream=stream,
+#         )
+
+#         llm_start_time = time.time()
+#         llm_chain = get_rag_llm_chain(**generate_input)
+#         llm_chain.invoke()
+
+#         answer = llm_generate(**generate_input)
+#         llm_end_time = time.time()
+#         elpase_time = llm_end_time - llm_start_time
+#         logger.info(f"runing time of llm: {elpase_time}s seconds")
+#         # answer = ret["answer"]
+#         debug_info["knowledge_qa_llm"] = answer
+#     except Exception as e:
+#         logger.info(f"Exception Query: {query_input}")
+#         logger.info(f"{traceback.format_exc()}")
+#         answer = ""
+
+#     # 5. update_session
+#     # start = time.time()
+#     # update_session(session_id=session_id, chat_session_table=chat_session_table,
+#     #                question=query_input, answer=answer, knowledge_sources=sources)
+#     # elpase_time = time.time() - start
+#     # logger.info(f'runing time of update_session : {elpase_time}s seconds')
+
+#     return answer, sources, contexts, debug_info
+
+
+# def get_strict_qq_chain(strict_q_q_index):
+#     def get_strict_qq_result(docs, threshold=0.7):
+#         results = []
+#         for doc in docs:
+#             if doc.metadata["score"] < threshold:
+#                 break 
+#             results.append(
+#                 {
+#                     "score": doc.metadata["score"],
+#                     "source": doc.metadata["source"],
+#                     "answer": doc.metadata["answer"],
+#                     "question": doc.metadata["question"],
+#                 }
+#             )
+#         return results
 
     mkt_q_q_retriever = QueryQuestionRetriever(
         index=strict_q_q_index,
@@ -589,495 +597,496 @@ def get_strict_qq_chain(strict_q_q_index):
     return strict_q_q_chain
 
 
-def return_strict_qq_result(x):
-    # def get_strict_qq_result(docs, threshold=0.7):
-    #     results = []
+# def return_strict_qq_result(x):
+#     # def get_strict_qq_result(docs, threshold=0.7):
+#     #     results = []
+#     #     for doc in docs:
+#     #         results.append({"score": doc.metadata["score"],
+#     #                         "source": doc.metadata["source"],
+#     #                         "answer": doc.metadata["answer"],
+#     #                         "question": doc.metadata["question"]})
+#     #     output = {"answer": json.dumps(results, ensure_ascii=False), "sources": [], "contexts": []}
+#     #     return output
+#     # return get_strict_qq_result(x["intent_info"]["strict_qq_intent_result"])
+#     return {
+#         "answer": json.dumps(
+#             x["intent_info"]["strict_qq_intent_result"], ensure_ascii=False
+#         ),
+#         "sources": [],
+#         "contexts": [],
+#         "context_docs": [],
+#         "context_sources": [],
+#     }
+
+
+# def get_rag_llm_chain(rag_config, stream):
+    # def contexts_trunc(docs: list, context_num=2):
+    #     # print('docs len',len(docs))
+    #     docs = [doc for doc in docs[:context_num]]
+    #     # the most related doc will be placed last
+    #     docs.sort(key=lambda x: x.metadata["score"])
+    #     # filter same docs
+    #     s = set()
+    #     context_strs = []
+    #     context_docs = []
+    #     context_sources = []
     #     for doc in docs:
-    #         results.append({"score": doc.metadata["score"],
-    #                         "source": doc.metadata["source"],
-    #                         "answer": doc.metadata["answer"],
-    #                         "question": doc.metadata["question"]})
-    #     output = {"answer": json.dumps(results, ensure_ascii=False), "sources": [], "contexts": []}
-    #     return output
-    # return get_strict_qq_result(x["intent_info"]["strict_qq_intent_result"])
-    return {
-        "answer": json.dumps(
-            x["intent_info"]["strict_qq_intent_result"], ensure_ascii=False
-        ),
-        "sources": [],
-        "contexts": [],
-        "context_docs": [],
-        "context_sources": [],
-    }
-
-
-def get_rag_llm_chain(rag_config, stream):
-    def contexts_trunc(docs: list, context_num=2):
-        # print('docs len',len(docs))
-        docs = [doc for doc in docs[:context_num]]
-        # the most related doc will be placed last
-        docs.sort(key=lambda x: x.metadata["score"])
-        # filter same docs
-        s = set()
-        context_strs = []
-        context_docs = []
-        context_sources = []
-        for doc in docs:
-            content = doc.page_content
-            if content not in s:
-                context_strs.append(content)
-                s.add(content)
-                context_docs.append({
-                    "doc": content,
-                    "source": doc.metadata["source"],
-                    "score": doc.metadata["score"]
-                    })
-                context_sources.append(doc.metadata["source"])
-        # print(len(context_docs))
-        # print(sg)
-        return {
-            "contexts": context_strs,
-            "context_docs": context_docs,
-            "context_sources":context_sources
-        }
+    #         content = doc.page_content
+    #         if content not in s:
+    #             context_strs.append(content)
+    #             s.add(content)
+    #             context_docs.append({
+    #                 "doc": content,
+    #                 "source": doc.metadata["source"],
+    #                 "score": doc.metadata["score"]
+    #                 })
+    #             context_sources.append(doc.metadata["source"])
+    #     # print(len(context_docs))
+    #     # print(sg)
+    #     return {
+    #         "contexts": context_strs,
+    #         "context_docs": context_docs,
+    #         "context_sources":context_sources
+    #     }
     
-    generator_llm_config = rag_config['generator_llm_config']
-    # TODO opt with efficiency
-    context_num = generator_llm_config['context_num']
-    contexts_trunc_stage = RunnablePassthrough.assign(
-        contexts=lambda x: contexts_trunc(x["docs"], context_num=context_num)['contexts'],
-        context_docs=lambda x: contexts_trunc(x["docs"], context_num=context_num)['context_docs'],
-        context_sources=lambda x: contexts_trunc(x["docs"], context_num=context_num)['context_sources'],
-    )
-    other_llm_config = copy.deepcopy(generator_llm_config)
-    other_llm_config.pop('model_id')
-    other_llm_config.pop('model_kwargs')
-    llm_chain = get_llm_chain(
-        model_id=generator_llm_config['model_id'],
-        intent_type=IntentType.KNOWLEDGE_QA.value,
-        model_kwargs=generator_llm_config['model_kwargs'],  # TODO
-        stream=stream,
-        # chat_history=rag_config['chat_history'],
-        **other_llm_config
-    )
-    llm_chain = contexts_trunc_stage |\
-          RunnablePassthrough.assign(chat_history=lambda x:rag_config['chat_history']) |\
-          RunnablePassthrough.assign(answer=llm_chain)
-    return llm_chain
-
-def get_qd_chain(
-    aos_index_list, retriever_top_k=10, reranker_top_k=5, using_whole_doc=True, chunk_num=0, enable_reranker=True
-):
-    retriever_list = [
-        QueryDocumentRetriever(
-            index, "vector_field", "text", "file_path", using_whole_doc, chunk_num, retriever_top_k, "zh", zh_embedding_endpoint
-        )
-        for index in aos_index_list
-    ] + [
-        QueryDocumentRetriever(
-            index, "vector_field", "text", "file_path", using_whole_doc, chunk_num, retriever_top_k, "en", en_embedding_endpoint
-        )
-        for index in aos_index_list
-    ]
-    lotr = MergerRetriever(retrievers=retriever_list)
-    if enable_reranker:
-        compressor = BGEReranker(top_n=reranker_top_k)
-    else:
-        compressor = MergeReranker(top_n=reranker_top_k)
-    compression_retriever = ContextualCompressionRetriever(
-        base_compressor=compressor, base_retriever=lotr
-    )
-    qd_chain = RunnablePassthrough.assign(docs=compression_retriever) 
-    return qd_chain
-
-def get_qd_llm_chain(
-    aos_index_list, 
-    rag_config, 
-    stream=False, 
-    # top_n=5
-):
-    using_whole_doc = rag_config['retriever_config']['using_whole_doc']
-    chunk_num = rag_config['retriever_config']['chunk_num']
-    retriever_top_k = rag_config['retriever_config']['retriever_top_k']
-    reranker_top_k = rag_config['retriever_config']['reranker_top_k']
-    enable_reranker = rag_config['retriever_config']['enable_reranker']
-    
-    llm_chain = get_rag_llm_chain(rag_config, stream)
-    qd_chain = get_qd_chain(aos_index_list, using_whole_doc=using_whole_doc,
-                            chunk_num=chunk_num, retriever_top_k=retriever_top_k,
-                            reranker_top_k=reranker_top_k, enable_reranker=enable_reranker)
-    qd_llm_chain = chain_logger(qd_chain, 'qd_retriever') | chain_logger(llm_chain,'llm_chain')
-    return qd_llm_chain
-
-
-def get_chat_llm_chain(
-        rag_config,
-        stream=False
-        ):
-    generator_llm_config = rag_config['generator_llm_config']
-    other_llm_config = copy.deepcopy(generator_llm_config)
-    other_llm_config.pop('model_id')
-    other_llm_config.pop('model_kwargs')
-
-    chat_llm_chain = get_llm_chain(
-        model_id=generator_llm_config['model_id'],
-        intent_type=IntentType.CHAT.value,
-        model_kwargs=generator_llm_config['model_kwargs'],  # TODO
-        stream=stream,
-        # chat_history=rag_config['chat_history'],
-        **other_llm_config
-    ) | {
-        "answer": lambda x: x,
-        "sources": lambda x: [],
-        "contexts": lambda x: [],
-        "intent_type": lambda x: IntentType.CHAT.value,
-        "context_docs": lambda x: [],
-        "context_sources": lambda x: [],
-    }
-    return chat_llm_chain
-
-def market_chain_entry(
-    query_input: str,
-    stream=False,
-    manual_input_intent=None,
-    rag_config=None
-):
-    """
-    Entry point for the Lambda function.
-
-    :param query_input: The query input.
-    :param aos_index: The index of the AOS engine.
-    :param stream(Bool): Whether to use llm stream decoding output.
-    return: answer(str)
-    """
-    assert rag_config is not None
-    generator_llm_config = rag_config['generator_llm_config']
-    intent_type = rag_config['intent_config']['intent_type']
-    aos_index_dict = json.loads(
-        os.environ.get(
-            "aos_index_dict",
-            '{"aos_index_mkt_qd":"aws-cn-mkt-knowledge","aos_index_mkt_qq":"gcr-mkt-qq","aos_index_dgr_qd":"ug-index","aos_index_dgr_qq":"faq-index-2"}',
-        )
-    )
-    aos_index_mkt_qd = aos_index_dict["aos_index_mkt_qd"]
-    aos_index_mkt_qq = aos_index_dict["aos_index_mkt_qq"]
-    aos_index_dgr_qd = aos_index_dict["aos_index_dgr_qd"]
-    aos_index_dgr_faq_qd = aos_index_dict["aos_index_dgr_faq_qd"]
-    aos_index_dgr_qq = aos_index_dict["aos_index_dgr_qq"]
-
-    # debug_info = {
-    #     "query": query_input,
-    #     "query_parser_info": {},
-    #     "q_q_match_info": {},
-    #     "knowledge_qa_knn_recall": {},
-    #     "knowledge_qa_boolean_recall": {},
-    #     "knowledge_qa_combined_recall": {},
-    #     "knowledge_qa_cross_model_sort": {},
-    #     "knowledge_qa_llm": {},
-    #     "knowledge_qa_rerank": {},
-    # }
-    debug_info = {}
-    contexts = []
-    sources = []
-    answer = ""
-    intent_info = {
-        "manual_input_intent": manual_input_intent,
-        "strict_qq_intent_result": {},
-    }
-
-    # 1. Strict Query Question Intent
-    # 1.1. strict query question retrieval.
-    # strict_q_q_chain = get_strict_qq_chain(aos_index_mkt_qq)
-
-    # 2. Knowledge QA Intent
-    # 2.1 query question retrieval.
-    dgr_q_q_retriever = QueryQuestionRetriever(
-        index=aos_index_dgr_qq,
-        vector_field="vector_field",
-        source_field="source",
-        size=5,
-        lang="zh",
-        embedding_model_endpoint=zh_embedding_endpoint
-    )
-    # 2.2 query document retrieval + LLM.
-    # qd_llm_chain = get_qd_llm_chain(
-    #     [aos_index_dgr_qd, aos_index_dgr_faq_qd, aos_index_mkt_qd],
-    #     rag_config,
-    #     stream,
-    #     # top_n=5,
-    #     # chunk_num=0
+    # generator_llm_config = rag_config['generator_llm_config']
+    # # TODO opt with efficiency
+    # context_num = generator_llm_config['context_num']
+    # RunnableDictAssign(contexts_trunc)
+    # contexts_trunc_stage = RunnablePassthrough.assign(
+    #     contexts=lambda x: contexts_trunc(x["docs"], context_num=context_num)['contexts'],
+    #     context_docs=lambda x: contexts_trunc(x["docs"], context_num=context_num)['context_docs'],
+    #     context_sources=lambda x: contexts_trunc(x["docs"], context_num=context_num)['context_sources'],
     # )
-
-    # 2.3 query question router.
-    def qq_route(info, threshold=0.9):
-        for doc in info["qq_result"]:
-            if doc.metadata["score"] > threshold:
-                output = {
-                    "answer": doc.metadata["answer"],
-                    "sources": doc.metadata["source"],
-                    "contexts": [],
-                    "context_docs": [],
-                    "context_sources": [],
-                    # "debug_info": lambda x: x["debug_info"],
-                }
-                logger.info('qq matched...')
-                info.update(output)
-                return info
-        qd_llm_chain = get_qd_llm_chain(
-            [aos_index_dgr_qd, aos_index_dgr_faq_qd, aos_index_mkt_qd],
-            rag_config,
-            stream,
-            # top_n=5,
-            # chunk_num=0
-        )
-        return qd_llm_chain
-
-    qq_chain = RunnablePassthrough.assign(qq_result=dgr_q_q_retriever)
-    qq_chain = chain_logger(qq_chain,'qq_chain')
-    qq_qd_llm_chain = qq_chain | RunnableLambda(qq_route)
-
-    # TODO design chat chain
     # other_llm_config = copy.deepcopy(generator_llm_config)
     # other_llm_config.pop('model_id')
     # other_llm_config.pop('model_kwargs')
-
-    
-    # chat_llm_chain = get_llm_chain(
+    # llm_chain = get_llm_chain(
     #     model_id=generator_llm_config['model_id'],
-    #     intent_type=IntentType.CHAT.value,
+    #     intent_type=IntentType.KNOWLEDGE_QA.value,
     #     model_kwargs=generator_llm_config['model_kwargs'],  # TODO
     #     stream=stream,
     #     # chat_history=rag_config['chat_history'],
     #     **other_llm_config
-    # ) | {
-    #     "answer": lambda x: x,
-    #     "sources": lambda x: [],
-    #     "contexts": lambda x: [],
-    #     "intent_type": lambda x: IntentType.CHAT.value,
-    #     "context_docs": lambda x: [],
-    #     "context_sources": lambda x: [],
-    # }
-    
-    # query process chain
-    query_process_chain = get_query_process_chain(
-        rag_config['chat_history'],
-        rag_config['query_process_config']
-    )
-    # | add_key_to_debug(add_key='conversation_query_rewrite',debug_key="debug_info")
-    #   | add_key_to_debug(add_key='query_rewrite',debug_key="debug_info")
-    
-    # query_rewrite_chain = chain_logger(
-    #     query_rewrite_chain,
-    #     "query rewrite module"
     # )
-    # intent recognition
-    intent_recognition_chain = auto_intention_recoginition_chain(
-        q_q_retriever_config={
-            "index_q_q":aos_index_mkt_qq,
-            'lang':'zh',
-            'embedding_endpoint':zh_embedding_endpoint,
-            "q_q_match_threshold": rag_config['retriever_config']['q_q_match_threshold']
-        },
-        intent_config=rag_config['intent_config']
-    )
+    # llm_chain = contexts_trunc_stage |\
+    #       RunnablePassthrough.assign(chat_history=lambda x:rag_config['chat_history']) |\
+    #       RunnablePassthrough.assign(answer=llm_chain)
+    # return llm_chain
 
-    intent_recognition_chain = chain_logger(
-        intent_recognition_chain,
-        'intention module',
-        log_output_template='intent chain output: {intent_type}'
-    )
+# def get_qd_chain(
+#     aos_index_list, retriever_top_k=10, reranker_top_k=5, using_whole_doc=True, chunk_num=0, enable_reranker=True
+# ):
+#     retriever_list = [
+#         QueryDocumentRetriever(
+#             index, "vector_field", "text", "file_path", using_whole_doc, chunk_num, retriever_top_k, "zh", zh_embedding_endpoint
+#         )
+#         for index in aos_index_list
+#     ] + [
+#         QueryDocumentRetriever(
+#             index, "vector_field", "text", "file_path", using_whole_doc, chunk_num, retriever_top_k, "en", en_embedding_endpoint
+#         )
+#         for index in aos_index_list
+#     ]
+#     lotr = MergerRetriever(retrievers=retriever_list)
+#     if enable_reranker:
+#         compressor = BGEReranker(top_n=reranker_top_k)
+#     else:
+#         compressor = MergeReranker(top_n=reranker_top_k)
+#     compression_retriever = ContextualCompressionRetriever(
+#         base_compressor=compressor, base_retriever=lotr
+#     )
+#     qd_chain = RunnablePassthrough.assign(docs=compression_retriever) 
+#     return qd_chain
+
+# def get_qd_llm_chain(
+#     aos_index_list, 
+#     rag_config, 
+#     stream=False, 
+#     # top_n=5
+# ):
+#     using_whole_doc = rag_config['retriever_config']['using_whole_doc']
+#     chunk_num = rag_config['retriever_config']['chunk_num']
+#     retriever_top_k = rag_config['retriever_config']['retriever_top_k']
+#     reranker_top_k = rag_config['retriever_config']['reranker_top_k']
+#     enable_reranker = rag_config['retriever_config']['enable_reranker']
+    
+#     llm_chain = get_rag_llm_chain(rag_config, stream)
+#     qd_chain = get_qd_chain(aos_index_list, using_whole_doc=using_whole_doc,
+#                             chunk_num=chunk_num, retriever_top_k=retriever_top_k,
+#                             reranker_top_k=reranker_top_k, enable_reranker=enable_reranker)
+#     qd_llm_chain = chain_logger(qd_chain, 'qd_retriever') | chain_logger(llm_chain,'llm_chain')
+#     return qd_llm_chain
+
+
+# def get_chat_llm_chain(
+#         rag_config,
+#         stream=False
+#         ):
+#     generator_llm_config = rag_config['generator_llm_config']
+#     other_llm_config = copy.deepcopy(generator_llm_config)
+#     other_llm_config.pop('model_id')
+#     other_llm_config.pop('model_kwargs')
+
+#     chat_llm_chain = get_llm_chain(
+#         model_id=generator_llm_config['model_id'],
+#         intent_type=IntentType.CHAT.value,
+#         model_kwargs=generator_llm_config['model_kwargs'],  # TODO
+#         stream=stream,
+#         # chat_history=rag_config['chat_history'],
+#         **other_llm_config
+#     ) | {
+#         "answer": lambda x: x,
+#         "sources": lambda x: [],
+#         "contexts": lambda x: [],
+#         "intent_type": lambda x: IntentType.CHAT.value,
+#         "context_docs": lambda x: [],
+#         "context_sources": lambda x: [],
+#     }
+#     return chat_llm_chain
+
+# def market_chain_entry(
+#     query_input: str,
+#     stream=False,
+#     manual_input_intent=None,
+#     rag_config=None
+# ):
+#     """
+#     Entry point for the Lambda function.
+
+#     :param query_input: The query input.
+#     :param aos_index: The index of the AOS engine.
+#     :param stream(Bool): Whether to use llm stream decoding output.
+#     return: answer(str)
+#     """
+#     assert rag_config is not None
+#     generator_llm_config = rag_config['generator_llm_config']
+#     intent_type = rag_config['intent_config']['intent_type']
+#     aos_index_dict = json.loads(
+#         os.environ.get(
+#             "aos_index_dict",
+#             '{"aos_index_mkt_qd":"aws-cn-mkt-knowledge","aos_index_mkt_qq":"gcr-mkt-qq","aos_index_dgr_qd":"ug-index","aos_index_dgr_qq":"faq-index-2"}',
+#         )
+#     )
+#     aos_index_mkt_qd = aos_index_dict["aos_index_mkt_qd"]
+#     aos_index_mkt_qq = aos_index_dict["aos_index_mkt_qq"]
+#     aos_index_dgr_qd = aos_index_dict["aos_index_dgr_qd"]
+#     aos_index_dgr_faq_qd = aos_index_dict["aos_index_dgr_faq_qd"]
+#     aos_index_dgr_qq = aos_index_dict["aos_index_dgr_qq"]
+
+#     # debug_info = {
+#     #     "query": query_input,
+#     #     "query_parser_info": {},
+#     #     "q_q_match_info": {},
+#     #     "knowledge_qa_knn_recall": {},
+#     #     "knowledge_qa_boolean_recall": {},
+#     #     "knowledge_qa_combined_recall": {},
+#     #     "knowledge_qa_cross_model_sort": {},
+#     #     "knowledge_qa_llm": {},
+#     #     "knowledge_qa_rerank": {},
+#     # }
+#     debug_info = {}
+#     contexts = []
+#     sources = []
+#     answer = ""
+#     intent_info = {
+#         "manual_input_intent": manual_input_intent,
+#         "strict_qq_intent_result": {},
+#     }
+
+#     # 1. Strict Query Question Intent
+#     # 1.1. strict query question retrieval.
+#     # strict_q_q_chain = get_strict_qq_chain(aos_index_mkt_qq)
+
+#     # 2. Knowledge QA Intent
+#     # 2.1 query question retrieval.
+#     dgr_q_q_retriever = QueryQuestionRetriever(
+#         index=aos_index_dgr_qq,
+#         vector_field="vector_field",
+#         source_field="source",
+#         size=5,
+#         lang="zh",
+#         embedding_model_endpoint=zh_embedding_endpoint
+#     )
+#     # 2.2 query document retrieval + LLM.
+#     # qd_llm_chain = get_qd_llm_chain(
+#     #     [aos_index_dgr_qd, aos_index_dgr_faq_qd, aos_index_mkt_qd],
+#     #     rag_config,
+#     #     stream,
+#     #     # top_n=5,
+#     #     # chunk_num=0
+#     # )
+
+#     # 2.3 query question router.
+#     def qq_route(info, threshold=0.9):
+#         for doc in info["qq_result"]:
+#             if doc.metadata["score"] > threshold:
+#                 output = {
+#                     "answer": doc.metadata["answer"],
+#                     "sources": doc.metadata["source"],
+#                     "contexts": [],
+#                     "context_docs": [],
+#                     "context_sources": [],
+#                     # "debug_info": lambda x: x["debug_info"],
+#                 }
+#                 logger.info('qq matched...')
+#                 info.update(output)
+#                 return info
+#         qd_llm_chain = get_qd_llm_chain(
+#             [aos_index_dgr_qd, aos_index_dgr_faq_qd, aos_index_mkt_qd],
+#             rag_config,
+#             stream,
+#             # top_n=5,
+#             # chunk_num=0
+#         )
+#         return qd_llm_chain
+
+#     qq_chain = RunnablePassthrough.assign(qq_result=dgr_q_q_retriever)
+#     qq_chain = chain_logger(qq_chain,'qq_chain')
+#     qq_qd_llm_chain = qq_chain | RunnableLambda(qq_route)
+
+#     # TODO design chat chain
+#     # other_llm_config = copy.deepcopy(generator_llm_config)
+#     # other_llm_config.pop('model_id')
+#     # other_llm_config.pop('model_kwargs')
+
+    
+#     # chat_llm_chain = get_llm_chain(
+#     #     model_id=generator_llm_config['model_id'],
+#     #     intent_type=IntentType.CHAT.value,
+#     #     model_kwargs=generator_llm_config['model_kwargs'],  # TODO
+#     #     stream=stream,
+#     #     # chat_history=rag_config['chat_history'],
+#     #     **other_llm_config
+#     # ) | {
+#     #     "answer": lambda x: x,
+#     #     "sources": lambda x: [],
+#     #     "contexts": lambda x: [],
+#     #     "intent_type": lambda x: IntentType.CHAT.value,
+#     #     "context_docs": lambda x: [],
+#     #     "context_sources": lambda x: [],
+#     # }
+    
+#     # query process chain
+#     query_process_chain = get_query_process_chain(
+#         rag_config['chat_history'],
+#         rag_config['query_process_config']
+#     )
+#     # | add_key_to_debug(add_key='conversation_query_rewrite',debug_key="debug_info")
+#     #   | add_key_to_debug(add_key='query_rewrite',debug_key="debug_info")
+    
+#     # query_rewrite_chain = chain_logger(
+#     #     query_rewrite_chain,
+#     #     "query rewrite module"
+#     # )
+#     # intent recognition
+#     intent_recognition_chain = auto_intention_recoginition_chain(
+#         q_q_retriever_config={
+#             "index_q_q":aos_index_mkt_qq,
+#             'lang':'zh',
+#             'embedding_endpoint':zh_embedding_endpoint,
+#             "q_q_match_threshold": rag_config['retriever_config']['q_q_match_threshold']
+#         },
+#         intent_config=rag_config['intent_config']
+#     )
+
+#     intent_recognition_chain = chain_logger(
+#         intent_recognition_chain,
+#         'intention module',
+#         log_output_template='intent chain output: {intent_type}'
+#     )
    
-    full_chain = query_process_chain | intent_recognition_chain  | RunnableBranch(
-        (lambda x:x['intent_type'] == IntentType.KNOWLEDGE_QA.value, qq_qd_llm_chain),
-        (lambda x:x['intent_type'] == IntentType.STRICT_QQ.value, return_strict_qq_result),
-        # (lambda x:x['intent_type'] == IntentType.STRICT_QQ.value, strict_q_q_chain),
-        get_chat_llm_chain(rag_config=rag_config,stream=stream),  # chat
-    )
-    # full_chain = intent_recognition_chain
-    # full_chain = RunnableLambda(route)
-    response = asyncio.run(full_chain.ainvoke(
-        {
-            "query": query_input,
-            "debug_info": debug_info,
-            "intent_type": intent_type,
-            "intent_info": intent_info,
-            "chat_history": rag_config['chat_history']
-        }
-    ))
+#     full_chain = query_process_chain | intent_recognition_chain  | RunnableBranch(
+#         (lambda x:x['intent_type'] == IntentType.KNOWLEDGE_QA.value, qq_qd_llm_chain),
+#         (lambda x:x['intent_type'] == IntentType.STRICT_QQ.value, return_strict_qq_result),
+#         # (lambda x:x['intent_type'] == IntentType.STRICT_QQ.value, strict_q_q_chain),
+#         get_chat_llm_chain(rag_config=rag_config,stream=stream),  # chat
+#     )
+#     # full_chain = intent_recognition_chain
+#     # full_chain = RunnableLambda(route)
+#     response = asyncio.run(full_chain.ainvoke(
+#         {
+#             "query": query_input,
+#             "debug_info": debug_info,
+#             "intent_type": intent_type,
+#             "intent_info": intent_info,
+#             "chat_history": rag_config['chat_history']
+#         }
+#     ))
 
-    answer = response["answer"]
-    sources = response["context_sources"]
-    contexts = response["context_docs"]
+#     answer = response["answer"]
+#     sources = response["context_sources"]
+#     contexts = response["context_docs"]
 
-    return answer, sources, contexts, debug_info
+#     return answer, sources, contexts, debug_info
 
-def market_conversation_summary_entry(
-        messages:list[dict],
-        rag_config=None,
-        stream=False
-    ):
+# def market_conversation_summary_entry(
+#         messages:list[dict],
+#         rag_config=None,
+#         stream=False
+#     ):
 
-    if not rag_config['chat_history']:
-        assert messages,messages
-        chat_history = []
-        for message in messages:
-            role = message['role']
-            content = message['content']
-            assert role in ['user','ai']
-            if role == 'user':
-                chat_history.append(HumanMessage(content=content))
-            else:
-                chat_history.append(AIMessage(content=content))
-        rag_config['chat_history'] = chat_history
+#     if not rag_config['chat_history']:
+#         assert messages,messages
+#         chat_history = []
+#         for message in messages:
+#             role = message['role']
+#             content = message['content']
+#             assert role in ['user','ai']
+#             if role == 'user':
+#                 chat_history.append(HumanMessage(content=content))
+#             else:
+#                 chat_history.append(AIMessage(content=content))
+#         rag_config['chat_history'] = chat_history
     
-    else:
-        # filter by the window time
-        time_window = rag_config.get('time_window',{})
-        start_time = time_window.get('start_time',-math.inf)
-        end_time = time_window.get('end_time',math.inf)
-        assert isinstance(start_time, float) and isinstance(end_time, float), (start_time, end_time)
-        chat_history = rag_config['chat_history']
-        chat_history = filter_chat_history_by_time(chat_history,start_time=start_time,end_time=end_time)
-        rag_config['chat_history'] = chat_history
-    # rag_config['intent_config']['intent_type'] = IntentType.CHAT.value
+#     else:
+#         # filter by the window time
+#         time_window = rag_config.get('time_window',{})
+#         start_time = time_window.get('start_time',-math.inf)
+#         end_time = time_window.get('end_time',math.inf)
+#         assert isinstance(start_time, float) and isinstance(end_time, float), (start_time, end_time)
+#         chat_history = rag_config['chat_history']
+#         chat_history = filter_chat_history_by_time(chat_history,start_time=start_time,end_time=end_time)
+#         rag_config['chat_history'] = chat_history
+#     # rag_config['intent_config']['intent_type'] = IntentType.CHAT.value
     
-    # query_input = """请简要总结上述对话中的内容,每一个对话单独一个总结，并用 '- '开头。 每一个总结要先说明问题。\n"""
-    mkt_conversation_summary_config = rag_config["mkt_conversation_summary_config"]
-    llm_chain = get_llm_chain(
-        intent_type=MKT_CONVERSATION_SUMMARY_TYPE,
-        stream=stream,
-        **mkt_conversation_summary_config, 
-    )
-    response = llm_chain.invoke({
-        "chat_history": rag_config['chat_history'],
-    })
-    return response, [], {}, {}
+#     # query_input = """请简要总结上述对话中的内容,每一个对话单独一个总结，并用 '- '开头。 每一个总结要先说明问题。\n"""
+#     mkt_conversation_summary_config = rag_config["mkt_conversation_summary_config"]
+#     llm_chain = get_llm_chain(
+#         intent_type=MKT_CONVERSATION_SUMMARY_TYPE,
+#         stream=stream,
+#         **mkt_conversation_summary_config, 
+#     )
+#     response = llm_chain.invoke({
+#         "chat_history": rag_config['chat_history'],
+#     })
+#     return response, [], {}, {}
 
-@timeit
-def main_qd_retriever_entry(
-    query_input: str,
-    aos_index: str,
-    rag_config=None,
-    manual_input_intent=None
-):
-    """
-    Entry point for the Lambda function.
+# @timeit
+# def main_qd_retriever_entry(
+#     query_input: str,
+#     aos_index: str,
+#     rag_config=None,
+#     manual_input_intent=None
+# ):
+#     """
+#     Entry point for the Lambda function.
 
-    :param query_input: The query input.
-    :param aos_index: The index of the AOS engine.
+#     :param query_input: The query input.
+#     :param aos_index: The index of the AOS engine.
 
-    return: answer(str)
-    """
-    debug_info = {
-        "query": query_input,
-        "query_parser_info": {},
-        "q_q_match_info": {},
-        "knowledge_qa_knn_recall": {},
-        "knowledge_qa_boolean_recall": {},
-        "knowledge_qa_combined_recall": {},
-        "knowledge_qa_cross_model_sort": {},
-        "knowledge_qa_llm": {},
-        "knowledge_qa_rerank": {},
-    }
-    retriever_top_k = rag_config['retriever_config']['retriever_top_k']
-    using_whole_doc = rag_config['retriever_config']['using_whole_doc']
-    chunk_num = rag_config['retriever_config']['chunk_num']
-    query_process_chain = get_query_process_chain(
-        rag_config['chat_history'],
-        rag_config['query_process_config']['query_rewrite_config'],
-        rag_config['query_process_config']['conversation_query_rewrite_config'],
-        rag_config['query_process_config']['hyde_config']
-    )
-    intent_type = rag_config['intent_config']['intent_type']
-    intent_info = {
-        "manual_input_intent": manual_input_intent,
-        "strict_qq_intent_result": {},
-    }
-    intent_recognition_chain = auto_intention_recoginition_chain("aos_index_mkt_qq")
-    intent_recognition_chain = chain_logger(
-        intent_recognition_chain,
-        'intention module',
-        log_output_template='intent chain output: {intent_type}'
+#     return: answer(str)
+#     """
+#     debug_info = {
+#         "query": query_input,
+#         "query_parser_info": {},
+#         "q_q_match_info": {},
+#         "knowledge_qa_knn_recall": {},
+#         "knowledge_qa_boolean_recall": {},
+#         "knowledge_qa_combined_recall": {},
+#         "knowledge_qa_cross_model_sort": {},
+#         "knowledge_qa_llm": {},
+#         "knowledge_qa_rerank": {},
+#     }
+#     retriever_top_k = rag_config['retriever_config']['retriever_top_k']
+#     using_whole_doc = rag_config['retriever_config']['using_whole_doc']
+#     chunk_num = rag_config['retriever_config']['chunk_num']
+#     query_process_chain = get_query_process_chain(
+#         rag_config['chat_history'],
+#         rag_config['query_process_config']['query_rewrite_config'],
+#         rag_config['query_process_config']['conversation_query_rewrite_config'],
+#         rag_config['query_process_config']['hyde_config']
+#     )
+#     intent_type = rag_config['intent_config']['intent_type']
+#     intent_info = {
+#         "manual_input_intent": manual_input_intent,
+#         "strict_qq_intent_result": {},
+#     }
+#     intent_recognition_chain = auto_intention_recoginition_chain("aos_index_mkt_qq")
+#     intent_recognition_chain = chain_logger(
+#         intent_recognition_chain,
+#         'intention module',
+#         log_output_template='intent chain output: {intent_type}'
         
-    )
-    qd_chain = get_qd_chain(
-        [aos_index], using_whole_doc=using_whole_doc, chunk_num=chunk_num, retriever_top_k=retriever_top_k, reranker_top_k=10
-    )
-    full_chain = query_process_chain | intent_recognition_chain | qd_chain
-    response = asyncio.run(full_chain.ainvoke({
-            "query": query_input,
-            "debug_info": debug_info,
-            "intent_type": intent_type,
-            "intent_info": intent_info,
-    }))
-    doc_list = []
-    for doc in response["docs"]:
-        doc_list.append({"page_content": doc.page_content, "metadata": doc.metadata})
-    return doc_list, debug_info
+#     )
+#     qd_chain = get_qd_chain(
+#         [aos_index], using_whole_doc=using_whole_doc, chunk_num=chunk_num, retriever_top_k=retriever_top_k, reranker_top_k=10
+#     )
+#     full_chain = query_process_chain | intent_recognition_chain | qd_chain
+#     response = asyncio.run(full_chain.ainvoke({
+#             "query": query_input,
+#             "debug_info": debug_info,
+#             "intent_type": intent_type,
+#             "intent_info": intent_info,
+#     }))
+#     doc_list = []
+#     for doc in response["docs"]:
+#         doc_list.append({"page_content": doc.page_content, "metadata": doc.metadata})
+#     return doc_list, debug_info
 
-def main_qq_retriever_entry(
-    query_input: str,
-    aos_index: str,
-):
-    """
-    Entry point for the Lambda function.
+# def main_qq_retriever_entry(
+#     query_input: str,
+#     aos_index: str,
+# ):
+#     """
+#     Entry point for the Lambda function.
 
-    :param query_input: The query input.
-    :param aos_index: The index of the AOS engine.
+#     :param query_input: The query input.
+#     :param aos_index: The index of the AOS engine.
 
-    return: answer(str)
-    """
-    debug_info = {
-        "query": query_input,
-        "query_parser_info": {},
-        "q_q_match_info": {},
-        "knowledge_qa_knn_recall": {},
-        "knowledge_qa_boolean_recall": {},
-        "knowledge_qa_combined_recall": {},
-        "knowledge_qa_cross_model_sort": {},
-        "knowledge_qa_llm": {},
-        "knowledge_qa_rerank": {},
-    }
-    full_chain = get_strict_qq_chain(aos_index)
-    response = full_chain.invoke({"query": query_input, "debug_info": debug_info})
-    return response
+#     return: answer(str)
+#     """
+#     debug_info = {
+#         "query": query_input,
+#         "query_parser_info": {},
+#         "q_q_match_info": {},
+#         "knowledge_qa_knn_recall": {},
+#         "knowledge_qa_boolean_recall": {},
+#         "knowledge_qa_combined_recall": {},
+#         "knowledge_qa_cross_model_sort": {},
+#         "knowledge_qa_llm": {},
+#         "knowledge_qa_rerank": {},
+#     }
+#     full_chain = get_strict_qq_chain(aos_index)
+#     response = full_chain.invoke({"query": query_input, "debug_info": debug_info})
+#     return response
 
-def main_chain_entry(
-        query_input: str,
-        aos_index: str,
-        stream=False,
-        rag_config=None
-):
-    """
-    Entry point for the Lambda function.
+# def main_chain_entry(
+#         query_input: str,
+#         aos_index: str,
+#         stream=False,
+#         rag_config=None
+# ):
+#     """
+#     Entry point for the Lambda function.
 
-    :param query_input: The query input.
-    :param aos_index: The index of the AOS engine.
+#     :param query_input: The query input.
+#     :param aos_index: The index of the AOS engine.
 
-    return: answer(str)
-    """
-    debug_info = {
-        "query": query_input,
-        "query_parser_info": {},
-        "q_q_match_info": {},
-        "knowledge_qa_knn_recall": {},
-        "knowledge_qa_boolean_recall": {},
-        "knowledge_qa_combined_recall": {},
-        "knowledge_qa_cross_model_sort": {},
-        "knowledge_qa_llm": {},
-        "knowledge_qa_rerank": {},
-    }
-    contexts = []
-    sources = []
-    answer = ""
-    full_chain = get_qd_llm_chain(
-        [aos_index], rag_config, stream
-    )
-    response = full_chain.invoke({"query": query_input, "debug_info": debug_info})
-    answer = response["answer"]
-    sources = response["context_sources"]
-    contexts = response["context_docs"]
-    return answer, sources, contexts, debug_info
+#     return: answer(str)
+#     """
+#     debug_info = {
+#         "query": query_input,
+#         "query_parser_info": {},
+#         "q_q_match_info": {},
+#         "knowledge_qa_knn_recall": {},
+#         "knowledge_qa_boolean_recall": {},
+#         "knowledge_qa_combined_recall": {},
+#         "knowledge_qa_cross_model_sort": {},
+#         "knowledge_qa_llm": {},
+#         "knowledge_qa_rerank": {},
+#     }
+#     contexts = []
+#     sources = []
+#     answer = ""
+#     full_chain = get_qd_llm_chain(
+#         [aos_index], rag_config, stream
+#     )
+#     response = full_chain.invoke({"query": query_input, "debug_info": debug_info})
+#     answer = response["answer"]
+#     sources = response["context_sources"]
+#     contexts = response["context_docs"]
+#     return answer, sources, contexts, debug_info
 
 def _is_websocket_request(event):
     """Check if the request is WebSocket or Restful
@@ -1094,17 +1103,17 @@ def _is_websocket_request(event):
     else:
         return False
 
-def get_retriever_response(docs, debug_info):
-    response = {"statusCode": 200, "headers": {"Content-Type": "application/json"}}
-    resp_header = {
-        "Content-Type": "application/json",
-        "Access-Control-Allow-Headers": "Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token",
-        "Access-Control-Allow-Origin": "*",
-        "Access-Control-Allow-Methods": "*",
-    }
-    response["body"] = json.dumps({"docs": docs, "debug_info": debug_info})
-    response["headers"] = resp_header
-    return response
+# def get_retriever_response(docs, debug_info):
+#     response = {"statusCode": 200, "headers": {"Content-Type": "application/json"}}
+#     resp_header = {
+#         "Content-Type": "application/json",
+#         "Access-Control-Allow-Headers": "Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token",
+#         "Access-Control-Allow-Origin": "*",
+#         "Access-Control-Allow-Methods": "*",
+#     }
+#     response["body"] = json.dumps({"docs": docs, "debug_info": debug_info})
+#     response["headers"] = resp_header
+#     return response
 
 # @handle_error
 def lambda_handler(event, context):
