+++ source/lambda/executor/utils/logger_utils.py
@@ -8,3 +8,5 @@ opensearch_logger = logging.getLogger("opensearch")
 opensearch_logger.setLevel(logging.ERROR)
 boto3_logger = logging.getLogger("botocore")
 boto3_logger.setLevel(logging.ERROR)
+
+
